# MP timing improvement